refactor code according to the comments
