+++ lib/Donors/DonorsContainer.js
@@ -1,4 +1,9 @@
-import { map, noop, sortBy } from 'lodash';
+import {
+  keyBy,
+  map,
+  noop,
+  sortBy,
+} from 'lodash';
 import PropTypes from 'prop-types';
 import { useMemo } from 'react';
 import { useIntl } from 'react-intl';
@@ -28,23 +33,18 @@ export function DonorsContainer({
   const intl = useIntl();
   const canViewOrganizations = stripes.hasPerm('ui-organizations.view');
 
-  const donorsMap = useMemo(() => {
-    return donors.reduce((acc, contact) => {
-      acc[contact.id] = contact;
-
-      return acc;
-    }, {});
-  }, [donors]);
+  const donorsMap = useMemo(() => keyBy(donors, 'id'), [donors]);
 
   const listOfDonors = useMemo(() => (fields.value || [])
-    .map((contactId, _index) => {
+    .reduce((acc, contactId, _index) => {
       const contact = donorsMap?.[contactId];
 
-      return {
-        ...(contact || { isDeleted: true }),
-        _index,
-      };
-    }), [donorsMap, fields.value]);
+      if (contact?.id) {
+        acc.push({ ...contact, _index });
+      }
+
+      return acc;
+    }, []), [donorsMap, fields.value]);
 
   const contentData = useMemo(() => sortBy(listOfDonors, [({ name }) => name?.toLowerCase()]), [listOfDonors]);
