+++ lib/PrivilegedDonorContacts/ContactsContainer.js
@@ -1,11 +1,15 @@
-import { map, sortBy } from 'lodash';
+import {
+  keyBy,
+  map,
+  sortBy,
+} from 'lodash';
 import PropTypes from 'prop-types';
 import { useMemo } from 'react';
 import { useIntl } from 'react-intl';
 
+import { useCategories } from '../hooks';
 import { acqRowFormatter } from '../utils';
 import { defaultContainerVisibleColumns } from './constants';
-import { useCategories } from './hooks';
 import { PrivilegedDonorContactsList } from './PrivilegedDonorContactsList';
 import { PrivilegedDonorContactsLookup } from './PrivilegedDonorContactsLookup';
 import {
@@ -13,7 +17,7 @@ import {
   getResultsFormatter,
 } from './utils';
 
-export function PrivilegedDonorContactsContainer({
+export function ContactsContainer({
   columnMapping,
   columnWidths,
   contacts,
@@ -23,7 +27,6 @@ export function PrivilegedDonorContactsContainer({
   orgId,
   setContactIds,
   searchLabel,
-  selectedContactIds,
   showTriggerButton,
   visibleColumns,
   ...rest
@@ -31,23 +34,18 @@ export function PrivilegedDonorContactsContainer({
   const intl = useIntl();
   const { categories } = useCategories();
 
-  const contactsMap = useMemo(() => {
-    return contacts.reduce((acc, contact) => {
-      acc[contact.id] = contact;
-
-      return acc;
-    }, {});
-  }, [contacts]);
+  const contactsMap = useMemo(() => keyBy(contacts, 'id'), [contacts]);
 
   const listOfDonors = useMemo(() => (fields.value || [])
-    .map((contactId, _index) => {
+    .reduce((acc, contactId, _index) => {
       const contact = contactsMap?.[contactId];
 
-      return {
-        ...(contact || { isDeleted: true }),
-        _index,
-      };
-    }), [contactsMap, fields.value]);
+      if (contact?.id) {
+        acc.push({ ...contact, _index });
+      }
+
+      return acc;
+    }, []), [contactsMap, fields.value]);
 
   const contentData = useMemo(() => sortBy(listOfDonors, [({ lastName }) => lastName?.toLowerCase()]), [listOfDonors]);
 
@@ -96,7 +94,6 @@ export function PrivilegedDonorContactsContainer({
             searchLabel={searchLabel}
             columnWidths={columnWidths}
             orgId={orgId}
-            selectedContactIds={selectedContactIds}
           />
         )
       }
@@ -104,7 +101,7 @@ export function PrivilegedDonorContactsContainer({
   );
 }
 
-PrivilegedDonorContactsContainer.propTypes = {
+ContactsContainer.propTypes = {
   columnMapping: PropTypes.object,
   columnWidths: PropTypes.object,
   contacts: PropTypes.arrayOf(PropTypes.object),
@@ -113,13 +110,12 @@ PrivilegedDonorContactsContainer.propTypes = {
   id: PropTypes.string,
   orgId: PropTypes.string,
   searchLabel: PropTypes.node,
-  selectedContactIds: PropTypes.arrayOf(PropTypes.string),
   setContactIds: PropTypes.func.isRequired,
   showTriggerButton: PropTypes.bool,
   visibleColumns: PropTypes.arrayOf(PropTypes.string),
 };
 
-PrivilegedDonorContactsContainer.defaultProps = {
+ContactsContainer.defaultProps = {
   showTriggerButton: true,
   visibleColumns: defaultContainerVisibleColumns,
 };
