+++ lib/PrivilegedDonorContacts/ContactsContainer.test.js
@@ -4,8 +4,8 @@ import user from '@testing-library/user-event';
 
 import stripesFinalForm from '@folio/stripes/final-form';
 
-import { PrivilegedDonorContactsContainer } from './PrivilegedDonorContactsContainer';
-import { useCategories } from './hooks';
+import { useCategories } from '../hooks';
+import { ContactsContainer } from './ContactsContainer';
 
 const mockVendor = { id: '1', name: 'Amazon' };
 
@@ -38,7 +38,7 @@ jest.mock('./PrivilegedDonorContactsLookup', () => ({
 
 const setContactIds = jest.fn();
 
-jest.mock('./hooks', () => ({
+jest.mock('../hooks', () => ({
   useCategories: jest.fn().mockReturnValue({
     categories: [],
     isLoading: false,
@@ -65,7 +65,7 @@ const defaultProps = {
 
 const renderForm = (props = {}) => (
   <form>
-    <PrivilegedDonorContactsContainer
+    <ContactsContainer
       {...defaultProps}
       {...props}
     />
