+++ lib/PrivilegedDonorContacts/PrivilegedDonorContacts.js
@@ -10,7 +10,7 @@ import {
 
 import { defaultColumnMapping } from './constants';
 import { useFetchPrivilegedContacts } from './hooks';
-import { PrivilegedDonorContactsContainer } from './PrivilegedDonorContactsContainer';
+import { ContactsContainer } from './ContactsContainer';
 
 export function PrivilegedDonorContacts({ name, privilegedContactIds, onChange, ...rest }) {
   const [contactIds, setContactIds] = useState(privilegedContactIds);
@@ -31,11 +31,10 @@ export function PrivilegedDonorContacts({ name, privilegedContactIds, onChange, 
         <FieldArray
           name={name}
           id={name}
-          component={PrivilegedDonorContactsContainer}
+          component={ContactsContainer}
           setContactIds={onSetContactIds}
           contacts={contacts}
           loading={isLoading}
-          selectedContactIds={contactIds}
           {...rest}
         />
       </Col>
