+++ lib/PrivilegedDonorContacts/PrivilegedDonorContacts.test.js
@@ -7,9 +7,12 @@ import stripesFinalForm from '@folio/stripes/final-form';
 import { useFetchPrivilegedContacts } from './hooks';
 import { PrivilegedDonorContacts } from './PrivilegedDonorContacts';
 
+jest.mock('../hooks', () => ({
+  useCategories: jest.fn().mockReturnValue({ categories: [] }),
+}));
+
 jest.mock('./hooks', () => ({
   useFetchPrivilegedContacts: jest.fn(),
-  useCategories: jest.fn().mockReturnValue({ categories: [] }),
 }));
 
 jest.mock('./PrivilegedDonorContactsLookup', () => ({
