+++ lib/PrivilegedDonorContacts/PrivilegedDonorContactsList.js
@@ -4,13 +4,13 @@ import { useIntl } from 'react-intl';
 
 import { MultiColumnList } from '@folio/stripes/components';
 
+import { useCategories } from '../hooks';
 import { acqRowFormatter } from '../utils';
 import {
   alignRowProps,
   defaultColumnMapping,
   defaultVisibleColumns,
 } from './constants';
-import { useCategories } from './hooks';
 import { getResultsFormatter } from './utils';
 
 export const PrivilegedDonorContactsList = ({
