+++ lib/PrivilegedDonorContacts/PrivilegedDonorContactsList.test.js
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 
 import { PrivilegedDonorContactsList } from './PrivilegedDonorContactsList';
 
-jest.mock('./hooks', () => ({
+jest.mock('../hooks', () => ({
   useCategories: jest.fn().mockReturnValue({
     categories: [],
     isLoading: false,
