+++ lib/PrivilegedDonorContacts/PrivilegedDonorContactsLookup.js
@@ -14,7 +14,6 @@ export const PrivilegedDonorContactsLookup = ({
   orgId,
   searchButtonStyle,
   searchLabel,
-  selectedContactIds,
 }) => {
   const stripes = useStripes();
 
@@ -44,7 +43,6 @@ export const PrivilegedDonorContactsLookup = ({
       addContacts={onAddContacts}
       renderNewContactBtn={renderNewContactBtn}
       isPrivilegedContactEnabled
-      selectedContactIds={selectedContactIds}
     >
       <span data-test-add-contact>
         <FormattedMessage id="stripes-acq-components.privilegedContacts.noFindContactPlugin" />
@@ -58,11 +56,9 @@ PrivilegedDonorContactsLookup.propTypes = {
   searchLabel: PropTypes.node,
   searchButtonStyle: PropTypes.string,
   orgId: PropTypes.string,
-  selectedContactIds: PropTypes.arrayOf(PropTypes.string),
 };
 
 PrivilegedDonorContactsLookup.defaultProps = {
   searchLabel: <FormattedMessage id="stripes-acq-components.privilegedContacts.button.addDonor" />,
   searchButtonStyle: 'default',
-  selectedContactIds: [],
 };
