+++ lib/PrivilegedDonorContacts/hooks/constants.js
@@ -1,13 +1,2 @@
 export const DEFAULT_DATA = [];
 export const PRIVILEGED_CONTACTS_API = 'organizations-storage/privileged-contacts';
-export const CATEGORIES_API = 'organizations-storage/categories';
-
-export const UNCATEGORIZED_VALUE = 'Uncategorized';
-
-export const VENDOR_DEFAULT_CATEGORIES = {
-  'Customer Service': 'customerService',
-  'Payments': 'payments',
-  'Returns': 'returns',
-  'Shipments': 'shipments',
-  [UNCATEGORIZED_VALUE]: 'uncategorized',
-};
