+++ lib/PrivilegedDonorContacts/hooks/index.js
@@ -1,3 +1,1 @@
 export { useFetchPrivilegedContacts } from './useFetchPrivilegedContacts';
-export { useCategories } from './useCategories';
-export { useTranslatedCategories } from './useTranslatedCategories';
