+++ lib/hooks/index.js
@@ -1,6 +1,7 @@
 export * from './useAccordionToggle';
 export * from './useAcqMethodsOptions';
 export * from './useAsyncDebounce';
+export * from './useCategories';
 export * from './useFunds';
 export * from './useIntegrationConfigs';
 export * from './useLineHoldings';
@@ -10,5 +11,6 @@ export * from './useOrganization';
 export * from './usePaneFocus';
 export * from './useShowCallout';
 export * from './useToggle';
+export * from './useTranslatedCategories';
 export * from './useUsersBatch';
 export * from './useLocalPagination';
