+++ lib/hooks/useCategories/constants.js
@@ -1,0 +1,1 @@
+export const CATEGORIES_API = 'organizations-storage/categories';
