+++ lib/hooks/useCategories/useCategories.js
@@ -5,8 +5,8 @@ import {
   useOkapiKy,
 } from '@folio/stripes/core';
 
-import { LIMIT_MAX } from '../../../constants';
-import { CATEGORIES_API } from '../constants';
+import { LIMIT_MAX } from '../../constants';
+import { CATEGORIES_API } from './constants';
 import { useTranslatedCategories } from '../useTranslatedCategories';
 
 const DEFAULT_DATA = [];
