+++ lib/hooks/useCategories/useCategories.test.js
