+++ lib/hooks/useTranslatedCategories/constants.js
@@ -1,0 +1,9 @@
+export const UNCATEGORIZED_VALUE = 'Uncategorized';
+
+export const VENDOR_DEFAULT_CATEGORIES = {
+  'Customer Service': 'customerService',
+  'Payments': 'payments',
+  'Returns': 'returns',
+  'Shipments': 'shipments',
+  [UNCATEGORIZED_VALUE]: 'uncategorized',
+};
