+++ lib/hooks/useTranslatedCategories/useTranslatedCategories.js
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { useIntl } from 'react-intl';
 
-import { VENDOR_DEFAULT_CATEGORIES } from '../constants';
+import { VENDOR_DEFAULT_CATEGORIES } from './constants';
 
 export const useTranslatedCategories = (categories) => {
   const intl = useIntl();
